+++ internal/mocks/doc.go
@@ -10,6 +10,7 @@ package mocks // import "github.com/slok/meterm/internal/mocks"
 //go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name Renderer
 //go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name GaugeWidget
 //go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name SinglestatWidget
+//go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name GraphWidget
 
 // Services mocks.
 //go:generate mockery -output ./service/metric -outpkg metric -dir ../service/metric -name Gatherer